Create API for Solver::solve().
The result of the solve method is a wrapper around a reference to the
solver. This means the solution can be expected without worrying the
solver state is changed, since no mutable reference to it can be given
out.

Later down the line, we probably want to implement the Drop trait for
SolveOutcome to reset the solver state. As such, dropping the
SolveOutcome allows for another call to solve, possibly with added
propagators or variables.